Adopt ruff as formatter and linter
Add ruff to the dev group and configure it in pyproject.toml at its default
line length and rule set, then format the codebase and clear every finding.

Two findings needed judgment rather than the autofix:

- atus_prediction.first_year_below_threshold wrapped round() in int(). round()
  on a float already returns an int, so the cast was a no-op; removed.
- test_ltt_extract builds its NAEP fixture as "\n".join([...]). FLY002 wants an
  f-string, but the list of lines is what makes the tabular layout under test
  legible, so it is suppressed with a noqa and a comment.

isort's known-first-party names the bare-imported pipeline modules (odds,
plotting, conftest, ...). They only resolve via the script-directory and
conftest sys.path entries, so without the list ruff files them under
third-party and merges them into the pandas/pytest block.

Also drops four noqa directives for rules ruff's default set does not enable.

Verified against the pipeline rather than tests alone: the 58 tests pass as
before, every data/derived/*.csv regenerates byte-identical, and
cited_values.py still matches each manuscript sentence. Figures are not
regenerated, since the scripts changed only in formatting and rewriting them
would churn 24 binaries on embedded timestamps alone.

+++ scripts/figures/atus_prediction.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from pathlib import Path
 import math
+from pathlib import Path
 from statistics import NormalDist, linear_regression
 
 import matplotlib.dates as mdates
@@ -12,7 +12,6 @@ import pandas as pd
 
 from plotting import save_figure
 from style import apply_style
-
 
 ACTIVITY = "Reading for personal interest"
 
@@ -82,9 +81,7 @@ def predict_with_pi(
     return y_hat, lower, upper
 
 
-def first_year_below_threshold(
-    model: dict[str, float], threshold: float
-) -> int | None:
+def first_year_below_threshold(model: dict[str, float], threshold: float) -> int | None:
     slope = model["slope"]
     intercept = model["intercept"]
     if slope == 0:
@@ -94,7 +91,7 @@ def first_year_below_threshold(
 
     crossing = (threshold - intercept) / slope
     if math.isclose(crossing, round(crossing), abs_tol=1e-9):
-        return int(round(crossing)) + 1
+        return round(crossing) + 1
     return math.ceil(crossing)
